refactor(order-srv|member-api)：用户端接口整改

+++ service-share/order-srv/src/main/java/com/lawu/eshop/order/srv/service/impl/ShoppingCartServiceImpl.java
@@ -7,15 +7,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.lawu.eshop.framework.web.ResultCode;
 import com.lawu.eshop.order.param.ShoppingCartSaveParam;
 import com.lawu.eshop.order.param.ShoppingCartUpdateParam;
 import com.lawu.eshop.order.srv.bo.ShoppingCartBO;
 import com.lawu.eshop.order.srv.converter.ShoppingCartConverter;
 import com.lawu.eshop.order.srv.domain.ShoppingCartDO;
 import com.lawu.eshop.order.srv.domain.ShoppingCartDOExample;
-import com.lawu.eshop.order.srv.domain.ShoppingOrderDO;
 import com.lawu.eshop.order.srv.mapper.ShoppingCartDOMapper;
-import com.lawu.eshop.order.srv.mapper.ShoppingOrderDOMapper;
 import com.lawu.eshop.order.srv.service.ShoppingCartService;
 
 /**
@@ -29,9 +28,6 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
 
 	@Autowired
 	private ShoppingCartDOMapper shoppingCartDOMapper;
-	
-	@Autowired
-	private ShoppingOrderDOMapper shoppingOrderDOMapper;
 
 	@Override
 	public List<ShoppingCartBO> findListByMemberId(Long memberId) {
@@ -44,43 +40,103 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
 		example.createCriteria().andMemberIdEqualTo(memberId);
 		return ShoppingCartConverter.convertBOS(shoppingCartDOMapper.selectByExample(example));
 	}
-	
+
+	/**
+	 * 加入购物车
+	 * 
+	 * @param memberId 会员id
+	 * @param param 保存参数
+	 * @return
+	 * @author Sunny
+	 */
 	@Transactional
 	@Override
-	public Long save(Long memberId, ShoppingCartSaveParam param) {
+	public int save(Long memberId, ShoppingCartSaveParam param) {
 		ShoppingCartDO suggestionDO = ShoppingCartConverter.convert(param);
 		suggestionDO.setMemberId(memberId);
 		suggestionDO.setGmtCreate(new Date());
 		suggestionDO.setGmtModified(new Date());
 
 		// 空值交给Mybatis去处理
-		Long result = (long) shoppingCartDOMapper.insertSelective(suggestionDO);
+		int result = shoppingCartDOMapper.insertSelective(suggestionDO);
 
-		// save返回id出去
-		return suggestionDO.getId() != null ? suggestionDO.getId() : result;
+		if (result <= 0) {
+			return ResultCode.SAVE_FAIL;
+		}
+
+		return ResultCode.SUCCESS;
 	}
-	
+
+	/**
+	 * 根据id更新购物车
+	 * 
+	 * @param id
+	 *            购物车id
+	 * @param memberId
+	 *            会员id
+	 * @param param
+	 *            更新参数
+	 * @return
+	 * @author Sunny
+	 */
 	@Transactional
 	@Override
-	public Integer update(Long id, ShoppingCartUpdateParam param) {
+	public int update(Long id, Long memberId, ShoppingCartUpdateParam param) {
+
+		ShoppingCartDO shoppingCartDO = shoppingCartDOMapper.selectByPrimaryKey(id);
+
+		if (shoppingCartDO == null || shoppingCartDO.getId() == null || shoppingCartDO.getId() <= 0) {
+			return ResultCode.RESOURCE_NOT_FOUND;
+		}
+
+		// 验证此条记录是否属于当前用户
+		if (!shoppingCartDO.getMemberId().equals(memberId)) {
+			return ResultCode.ILLEGAL_OPERATION;
+		}
+
 		ShoppingCartDO suggestionDO = ShoppingCartConverter.convert(param);
 		suggestionDO.setId(id);
 		suggestionDO.setGmtModified(new Date());
-		
+
 		// 空值交给Mybatis去处理
-		return shoppingCartDOMapper.updateByPrimaryKeySelective(suggestionDO);
+		shoppingCartDOMapper.updateByPrimaryKeySelective(suggestionDO);
+
+		return ResultCode.SUCCESS;
 	}
 
+	/**
+	 * 根据id删除购物车的商品
+	 * 
+	 * @param id
+	 *            购物车id
+	 * @param memberId
+	 *            会员id
+	 * @return
+	 * @author Sunny
+	 */
 	@Transactional
 	@Override
-	public Integer remove(Long id) {
-		return shoppingCartDOMapper.deleteByPrimaryKey(id);
+	public int remove(Long id, Long memberId) {
+		ShoppingCartDO shoppingCartDO = shoppingCartDOMapper.selectByPrimaryKey(id);
+
+		if (shoppingCartDO == null || shoppingCartDO.getId() == null || shoppingCartDO.getId() <= 0) {
+			return ResultCode.RESOURCE_NOT_FOUND;
+		}
+
+		if (!shoppingCartDO.getMemberId().equals(memberId)) {
+			return ResultCode.ILLEGAL_OPERATION;
+		}
+
+		shoppingCartDOMapper.deleteByPrimaryKey(id);
+
+		return ResultCode.SUCCESS;
 	}
-	
+
 	/**
 	 * 根据购物车id列表查询购物车列表
 	 * 
-	 * @param ids 购物车id列表
+	 * @param ids
+	 *            购物车id列表
 	 * @return
 	 */
 	@Override
@@ -88,27 +144,15 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
 		if (ids == null || ids.isEmpty()) {
 			return null;
 		}
-		
+
 		ShoppingCartDOExample example = new ShoppingCartDOExample();
 		example.createCriteria().andIdIn(ids);
-		
+
 		return ShoppingCartConverter.convertBOS(shoppingCartDOMapper.selectByExample(example));
 	}
-	
-	public ShoppingCartDO get(Long id){
+
+	public ShoppingCartDO get(Long id) {
 		return shoppingCartDOMapper.selectByPrimaryKey(id);
 	}
 
-	@Override
-	public Integer deleteByShoppingOrderId(Long shoppingOrderId) {
-		if (shoppingOrderId == null || shoppingOrderId <= 0) {
-			return null;
-		}
-		
-		ShoppingOrderDO shoppingOrderDO = shoppingOrderDOMapper.selectByPrimaryKey(shoppingOrderId);
-		String shoppingCartIdsStr = shoppingOrderDO.getShoppingCartIdsStr();
-		
-		
-		return null;
-	}
 }
